requirements.txt
pandas
openpyxl
PyYaml
openai
jsonlines
requests
boto3